Planner: move tests to submodules
Move the 57 test functions from planner/mod.rs into their respective
submodule files: resolver.rs (23 tests), select.rs (27 tests), and
dml.rs (7 tests). Each submodule now owns the tests for its own logic.
